+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes in 11.6
 - CSS Vertical align button inside first td, use when row has input (list with fields) in stylesheet.css
 - CSS WPadmin theme use OpenSans Semibold font instead of OpenSans Bold in font.css
 - Add vertically align list data option (defaults to false) in ListOutput.fnc.php
+- CSS add .has-input class to list having input (input or select fields) in ListOutput.fnc.php
 
 Changes in 11.5.1
 -----------------